// sitat-fe-workspace/projects/app-contratti/src/app/modules/models/fasi/fase-aggiudicazione.model.ts
export interface FaseAggiudicazioneEntry {
    altreSomme?: number;
    astaElettronica?: string;
    codGara?: number;
    codLotto?: number;
    codStrumento?: string;
    counter?: number;
    dataAtto?: Date;
    dataInvito?: Date;
    dataManifestInteresse?: Date;
    dataScadRichiestaInvito?: Date;
    dataScadenzaPresOfferta?: Date;
    dataVerbAggiudicazione?: Date;
    flagImporti?: string;
    flagLivelloProgettuale?: string;
    flagRichSubappalto?: string;
    flagSicurezza?: string;
    flagSubappFinReg?: string;
    idIndizione?: number;
    impEsclusInsuffGiust?: number;
    impNonAssog?: number;
    importoAggiudicazione?: number;
    importoComplAppalto?: number;
    importoComplIntervento?: number;
    importoDisposizione?: number;
    importoForniture?: number;
    importoIva?: number;
    importoLavori?: number;
    importoProgettazione?: number;
    importoServizi?: number;
    importoSicurezza?: number;
    importoSommeOpzioniRinnovi?: number;
    importoSommeripetizioni?: number;
    importosubtotale?: number;
    iva?: number;
    modalitaRiaggiudicazione?: number;
    numImpreseInvitate?: number;
    numImpreseOfferenti?: number;
    numImpreseRichiedenti?: number;
    numManifestInteresse?: number;
    numOfferteAmmesse?: number;
    numOfferteEscluse?: number;
    numOfferteFuoriSoglia?: number;
    numeroAtto?: string;
    offertaMassima?: number;
    offertaMinima?: number;
    opereUrbanizScomputo?: string;
    percOffAumento?: number;
    percentRibassoAgg?: number;
    preinformazione?: string;
    proceduraAcc?: string;
    pubblicata?: boolean;
    requisitiSettSpec?: number;
    termineRidotto?: string;
    tipoAtto?: number;
    valoreSogliaAnomalia?: number;
    verificaCampione?: string;
}

export interface FaseAggInsertForm {
    altreSomme?: number;
    astaElettronica?: string;
    codGara?: number;
    codLotto?: number;
    codStrumento?: string;
    dataAtto?: Date;
    dataInvito?: Date;
    dataManifestInteresse?: Date;
    dataScadRichiestaInvito?: Date;
    dataScadenzaPresOfferta?: Date;
    dataVerbAggiudicazione?: Date;
    flagImporti?: string;
    flagLivelloProgettuale?: string;
    flagRichSubappalto?: string;
    flagSicurezza?: string;
    flagSubappFinReg?: string;
    idIndizione?: number;
    impEsclusInsuffGiust?: number;
    impNonAssog?: number;
    importoAggiudicazione?: number;
    importoComplAppalto?: number;
    importoComplIntervento?: number;
    importoDisposizione?: number;
    importoForniture?: number;
    importoIva?: number;
    importoLavori?: number;
    importoProgettazione?: number;
    importoServizi?: number;
    importoSicurezza?: number;
    importoSommeOpzioniRinnovi?: number;
    importoSommeripetizioni?: number;
    importosubtotale?: number;
    iva?: number;
    modalitaRiaggiudicazione?: number;
    num?: number;
    numImpreseInvitate?: number;
    numImpreseOfferenti?: number;
    numImpreseRichiedenti?: number;
    numManifestInteresse?: number;
    numOfferteAmmesse?: number;
    numOfferteEscluse?: number;
    numOfferteFuoriSoglia?: number;
    numeroAtto?: string;
    offertaMassima?: number;
    offertaMinima?: number;
    opereUrbanizScomputo?: string;
    percOffAumento?: number;
    percentRibassoAgg?: number;
    preinformazione?: string;
    proceduraAcc?: string;
    requisitiSettSpec?: number;
    termineRidotto?: string;
    tipoAtto?: number;
    valoreSogliaAnomalia?: number;
    verificaCampione?: string;
}